Rename mod mongo to mongodb
